Implement repository interfaces for User, RefreshToken, and Book repositories
